# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toya Amerta API",
        "description": "REST API untuk aplikasi mobile (Flutter) Sistem Informasi PDAM Desa Toya Amerta. Semua response menggunakan envelope { success, message, data, meta }.",
        "contact": {
            "name": "Toya Amerta Dev Team",
            "email": "pt.andisuartika@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/petugas/customers": {
            "get": {
                "tags": [
                    "Petugas - Catat Meter"
                ],
                "summary": "Daftar pelanggan aktif",
                "description": "Digunakan untuk mengisi dropdown pelanggan saat input meter.",
                "operationId": "7beb832231a9d00d0677a407a90d206e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "customer_number": {
                                                        "type": "string",
                                                        "example": "PLG-0005"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Wayan Karya"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Banjar Kaja No. 12"
                                                    },
                                                    "zone": {
                                                        "type": "string",
                                                        "example": "Zona A"
                                                    },
                                                    "tariff": {
                                                        "type": "string",
                                                        "example": "Tarif Rumah Tangga"
                                                    },
                                                    "initial_meter": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 100
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 45
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/petugas/water-readings/{id}": {
            "get": {
                "tags": [
                    "Petugas - Catat Meter"
                ],
                "summary": "Detail pembacaan meter",
                "operationId": "9ce15a44630b021cf0902c3680a7e7ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pembacaan meter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WaterReading"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UnauthorizedResponse": {
                "title": "Unauthenticated Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated. Token tidak valid atau sudah kedaluwarsa."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ForbiddenResponse": {
                "title": "Forbidden Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Akses ditolak. Role Anda tidak memiliki izin."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "WaterReading": {
                "title": "Water Reading",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 55
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "Wayan Karya"
                    },
                    "customer_number": {
                        "type": "string",
                        "example": "PLG-0005"
                    },
                    "zone": {
                        "type": "string",
                        "example": "Zona A"
                    },
                    "officer_name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "period_year": {
                        "type": "integer",
                        "example": 2026
                    },
                    "period_month": {
                        "type": "integer",
                        "example": 6
                    },
                    "period_label": {
                        "type": "string",
                        "example": "Juni 2026"
                    },
                    "reading_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-03"
                    },
                    "previous_reading": {
                        "type": "number",
                        "format": "float",
                        "example": 132.8
                    },
                    "current_reading": {
                        "type": "number",
                        "format": "float",
                        "example": 145.2
                    },
                    "usage_m3": {
                        "type": "number",
                        "format": "float",
                        "example": 12.4
                    },
                    "price_per_m3": {
                        "type": "number",
                        "format": "float",
                        "example": 2500
                    },
                    "minimum_charge": {
                        "type": "number",
                        "format": "float",
                        "example": 15000
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 31000
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "belum_bayar",
                        "enum": [
                            "belum_bayar",
                            "sebagian",
                            "lunas"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Meteran normal",
                        "nullable": true
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://your-domain.com/storage/water-readings/abc123.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token yang didapat dari endpoint /auth/login. Format header: Authorization: Bearer {token}",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}